Move streaming service files to a separate folder
