+++ src/modules/history/components/history/HistoryList.tsx
@@ -2,7 +2,7 @@ import { List } from '@material-ui/core';
 import * as React from 'react';
 import { WrongItemDialog } from '../../../../components/WrongItemDialog';
 import { Item } from '../../../../models/Item';
-import { StreamingServiceId } from '../../../../streaming-services';
+import { StreamingServiceId } from '../../../../streaming-services/streaming-services';
 import { HistoryListItem } from './HistoryListItem';
 
 interface HistoryListProps {
