+++ src/modules/history/components/history/HistoryOptionsList.tsx
@@ -1,7 +1,7 @@
 import { Button, ButtonGroup, FormGroup, Typography } from '@material-ui/core';
 import * as React from 'react';
 import { SyncOption } from '../../../../services/BrowserStorage';
-import { Store } from '../../streaming-services/common/Store';
+import { Store } from '../../../../streaming-services/common/Store';
 import { HistoryOptionsListItem } from './HistoryOptionsListItem';
 
 interface HistoryOptionsListProps {
